# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRUD Laravel ",
        "description": "API para administrar multimples servicios y relaciones entre tablas",
        "contact": {
            "email": "jhonfredyaya04@gmail.com"
        },
        "license": {
            "name": "Repocitorio",
            "url": "https://github.com/JhonFac/Crud_Laravel.git"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/company/page/{id}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Obtiene todos los companies",
                "description": "Display a listing of the resource.",
                "operationId": "d34aadbac2cb8d15f0f5cc9dfa15f95d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Numero de la pagina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de companies"
                    }
                }
            }
        },
        "/api/company": {
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Crea un nuevo company",
                "description": "Store a newly created resource in storage.",
                "operationId": "faf1545d2af89079ca55c4f9ed6ea302",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "image_path": {
                                        "type": "string"
                                    },
                                    "industry": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "company creado exitosamente"
                    }
                }
            }
        },
        "/api/company/{id}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Obtiene el company por ID",
                "description": "Display the specified resource.",
                "operationId": "29b0f8f11c5e74b0c7a9f94f223c0abc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Numero de la pagina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de companies"
                    }
                }
            },
            "put": {
                "tags": [
                    "Company"
                ],
                "summary": "Actualiza un company existente",
                "description": "Update the specified resource in storage.",
                "operationId": "dba5691bd7818c368accfd785048ea90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del company a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "image_path": {
                                        "type": "string"
                                    },
                                    "industry": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "company actualizado exitosamente"
                    },
                    "404": {
                        "description": "company no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Company"
                ],
                "summary": "Elimina un company existente",
                "description": "Remove the specified resource from storage.",
                "operationId": "a073402836212623826fbe1e5a591454",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del company a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "company eliminado exitosamente"
                    },
                    "404": {
                        "description": "company no encontrado"
                    }
                }
            }
        },
...
    },
...
}